ath9k_htc: add PS support
Power Save support for ath9k_htc is centered around waking up the card
for RX traffic. This is achieved using hardware timers that are armed
each time we receive a beacon from one of the APs.

Limitations:
- 2 vifs;
- same channel;
- DTIM count = 2 for both APs.

Bugs:

1. there are cases when the NULL function is not ACK'ed for the second
vif and this prevents the wifi card from entering PS mode. After PS
is enabled (iw dev wlanX set power save on), make sure that /var/log/dmesg
contains 'sdata2' string (or check ieee80211_tx_status in case the
corresponding printk for 'sdata2' changes).

2. The WiFi card should wake up at each 200ms for AP1 and each 200ms
for AP2 (for catching beacons with the DTIM bit set). DTIM, TIM and
nextbtt hardware timers are armed with these values but there are cases
when we miss some beacons. That's because we don't have a good estimation
for the TSF of the last received beacon. Maybe we could use a register
from AR9271 that holds the TSF of the last received becon?

3. After a time, the Power Monitor device shows that the card exists PS
Mode but the dmesg logs shows that the card still captures traffic
according to the hardware timers (e.g.: there are quiet periods when no
data is received, still the power consumption of the card is high, as if
PS mode would be disabled).

More details can be found at:
https://github.com/doru91/linux-stable/wiki/Two-Interfaces-Power-Save

Signed-off-by: Doru Gucea <gucea.doru@gmail.com>